chore(rust): move `error.rs` of `ockam_identity` one level up and remove unused errors
