+++ implementations/rust/ockam/ockam_identity/src/identity/mod.rs
@@ -1,5 +1,3 @@
-mod error;
-
 #[allow(clippy::module_inception)]
 mod identity;
 /// List of key changes associated to an identity
@@ -7,7 +5,6 @@ pub mod identity_change;
 mod identity_change_history;
 mod identity_identifier;
 
-pub use error::*;
 pub use identity::*;
 pub use identity_change::*;
 pub use identity_change_history::*;
